Quiet various warnings.

+++ Sparkle/SUSystemProfiler.m
@@ -165,7 +165,7 @@ NSString *const SUSystemProfilerPreferredLanguageKey = @"lang";
     size_t bytes_size = sizeof(unsigned long);
     if (sysctlbyname("hw.memsize", &bytes, &bytes_size, NULL, 0) == 0) {
         double megabytes = bytes / (1024. * 1024.);
-        NSString *stringValue = [NSString stringWithFormat:@"%lu", (unsigned long)megabytes];
+        NSString *stringValue = [NSString stringWithFormat:@"%f", megabytes];
         [profileArray addObject:[NSDictionary dictionaryWithObjects:@[SUSystemProfilerMemoryKey, @"Memory (MB)", stringValue, stringValue] forKeys:profileDictKeys]];
     }
